Rename const.py to constants.py
