+++ conda_verify/constants.py
@@ -1,16 +1,5 @@
-LICENSE_FAMILIES = set("""
-AGPL
-GPL2
-GPL3
-LGPL
-BSD
-MIT
-Apache
-PSF
-Public-Domain
-Proprietary
-Other
-""".split())
+LICENSE_FAMILIES = ['AGPL', 'GPL2', 'GPL3', 'LGPL', 'BSD', 'MIT', 'Apache',
+                    'PSF', 'Public-Domain', 'Proprietary', 'Other']
 
 FIELDS = {
     'package': {'name', 'version'},
